Refactor load aggregator with 'stress_test' module
This makes the core more readable and usable.
